% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bioassay}
\alias{bioassay}
\title{Example bioassy data set}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(bioassay)
}
\description{
Example bioassy data set
}
\examples{
data(bioassay)
head(bioassay$assay1)
}
\keyword{datasets}
