% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratios.R
\name{model.signif}
\alias{model.signif}
\title{Test mortality model significance}
\usage{
model.signif(data)
}
\arguments{
\item{data}{a data frame of probit transformed mortality data using the function probit.trans}
}
\value{
a list with model outputs: a chi-square test if there only two strains or bonferroni test significance from a pariwise model comparison if there are more than two strains
}
\description{
This function tests whether there are differences between the regression of mortality between the strains (likelihood ratio test); It compares a model with and without different strains (i.e. to compare the significance if the data came from one strain or different strains)
}
\examples{
data(bioassay)
transd<-probid.trans(bioassay$assay2)
data<-transd$tr.data
model.signif(data)

}
\author{
Pascal Milesi, Piyal Karunarathne
}
