% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mort.plot}
\alias{mort.plot}
\title{Plot mortality}
\usage{
mort.plot(
  data,
  strains = NULL,
  plot.conf = TRUE,
  conf.level = 0.95,
  test.validity = TRUE,
  ...
)
}
\arguments{
\item{data}{a data frame of probit transformed mortality data using the function probit.trans}

\item{strains}{character. list of test strains. If not provided, the function will plot all the strains in the data set.}

\item{plot.conf}{logical. Whether to plot the confidence level on the plot for each strain}

\item{conf.level}{numerical. The confidence interval to be plotted}

\item{test.validity}{logical. Test and plot the validity of the models}

\item{...}{parameters to be passed on to graphics for the plot (e.g. col, pch)}
}
\value{
A plot of mortality
}
\description{
Plots the mortality of a given data set with and without confidence levels, and adds the validity of the regression.
}
\examples{
data(bioassay)
transd<-probid.trans(bioassay$assay2)
data<-transd$tr.data
strains<-levels(data$strain)
mort.plot(data,strains)

}
\author{
Piyal Karunarathne, Pascal Milesi
}
