% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratios.R
\name{resist.ratio}
\alias{resist.ratio}
\title{Calculate regression coefficients and resistance ratios}
\usage{
resist.ratio(
  data,
  conf.level = 0.95,
  ref.strain = NULL,
  plot = FALSE,
  plot.conf = TRUE,
  test.validity = TRUE,
  ...
)
}
\arguments{
\item{data}{a data frame of probit transformed mortality data using the function probit.trans}

\item{conf.level}{numerical. confidence level for the LD calculation (default 0.95)}

\item{ref.strain}{character. name of the reference strain if present (see details)}

\item{plot}{logical. Whether to plot the mortality plot}

\item{plot.conf}{logical. If plot=TRUE, whether to plot the confidence intervals}

\item{test.validity}{logical. (If plot=TRUE) Test and plot the validity of the models}

\item{...}{parameters to be passed on to graphics for the plot (e.g. col, pch)}
}
\value{
returns a data frame of statistical output and if plot=TRUE, plots the mortality on a log transformed scale.
}
\description{
This function calculates coefficients of the linear regression and the LD values for the strains tested using a GLM. Further it also outputs 95 percent confidence intervals of the RR (log-dose), according to Robertson and Preisler 1992 (also check references).
}
\details{
If a name of the reference strain is provided as it appears in the data, it will be used in the model. If not provided, the function will look for strains with the suffix "ref" in their names and will use it as the reference. If this returns NULL, the function will run the model considering no reference strain.
}
\examples{
data(bioassay)
transd<-probid.trans(bioassay$assay2)
data<-transd$tr.data
resist.ratio(data,plot=TRUE)

}
\references{
Finney DJ(1971). Probitanalysis. Cambridge:Cambridge UniversityPress. 350p.

HommelG(1988). A stage wise rejective multiple test procedure based on a modified Bonferroni test. Biometrika 75, 383-6.

Johnson RM, Dahlgren L, Siegfried BD,EllisMD(2013). Acaricide,fungicide and druginteractions in honeybees (Apis mellifera). PLoSONE8(1): e54092.

Robertson, J. L., and H.K. Preisler.1992. Pesticide bioassays with arthropods. CRC, Boca Raton, FL.
}
\author{
Pascal Milesi, Piyal Karunarathne
}
